requirements.txt
aiohttp==3.8.4
colorama==0.4.6
pandas==1.5.3
pyfiglet==1.0.2
python-dateutil==2.8.2